backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.19
pymysql==1.1.0
pydantic==2.0.3
python-dotenv==1.0.0
web3==6.5.0
python-multipart==0.0.6

